fix(editor): revert task list dependence on ids
This partially reverts 5b2723665273aaa2035706e8720ad736399130e0. Adding ids to task list items is not as simple and actually made it worse in some cases. Hence we stick to comparing the content of nodes for now, until this is properly fixed in tiptap.

Related https://kolaente.dev/vikunja/vikunja/issues/2091#issuecomment-60063