fix(tasks): clarify usage of repeating modes available in quick add magic.
Originally, it was possible to specify repeating intervals for months or years. This was actually misleading because the interval would be converted to seconds under the hood, causing all kinds of problems with leap years and months with 30 vs 31 days. Because of that, the options for month and year were removed in 1cc14a06a1895dd5c89b546514ae742f2e1c0888 to make it clear what these actually meant.
Now, since we do have a repeating mode "Monthly", this commit changes the behaviour to repeat a task monthly when it was specified as "every month", but only that. All other cases were removed from the docs since they are not actually supported.

https://github.com/go-vikunja/vikunja/issues/271