fix(kanban): set new bucket id on task after moving it
This fixes a bug where a bucket update in the api was not triggered because the task still had the old bucket id saved. This led to inconsistencies between what was visible in the UI and what was actually stored.

Probably introduced in 12c0311ccb4c78e16c903c5bbc49f818cf08da32

Resolves https://community.vikunja.io/t/warning-when-moving-tasks-to-a-bucket-with-limit/3520